Added the new user interface
This user interface offers new features that are more aligned with what we expected from this project when we first decided to work on. Also, it is more user-friendly and ergonomic. The user interface has a minimalistic design that offers necessary functionalities without too much cognitive overload for the user. The reasons behind this approach was more or less these articles:
https://en.wikipedia.org/wiki/Chunking_(psychology)
https://en.wikipedia.org/wiki/The_Magical_Number_Seven,_Plus_or_Minus_Two